fix: compass
