+++ client/src/ui/utils/Camera.tsx
@@ -8,6 +8,7 @@ import { useRoute } from "wouter";
 import { soundSelector, useUiSounds } from "../../hooks/useUISound";
 import * as THREE from "three";
 import useUIStore from "@/hooks/store/useUIStore";
+import { throttle } from "lodash";
 
 interface Props {
   position: {
@@ -35,8 +36,8 @@ const minPan = new THREE.Vector3(0, -Infinity, -1400);
 const maxPan = new THREE.Vector3(2700, Infinity, 0);
 
 const CameraControls = ({ position, target }: Props) => {
-  const setCameraPosition = useUIStore((state) => state.setCameraPosition);
-  const setCameraTarget = useUIStore((state) => state.setCameraTarget);
+  const direction = useUIStore((state) => state.compassDirection);
+  const setCompassDirection = useUIStore((state) => state.setCompassDirection);
 
   const {
     camera,
@@ -98,15 +99,39 @@ const CameraControls = ({ position, target }: Props) => {
     playFly();
   }, [target, position]);
 
+  const updateCompassDirection = useCallback(
+    throttle((cameraPosition, cameraTarget) => {
+      const { x: posX, z: posZ } = cameraPosition;
+      const { x: targetX, z: targetZ } = cameraTarget;
+
+      const dx = targetX - posX;
+      const dz = targetZ - posZ;
+      const angleRadians = Math.atan2(dz, dx);
+      const angleDegrees = angleRadians * (180 / Math.PI);
+      const normalizedNewAngle = (angleDegrees + 360) % 360;
+
+      // Determine the shortest path between the old and new angles
+      const deltaAngle = normalizedNewAngle - direction;
+      const shortestDelta = (deltaAngle + 360) % 360;
+      const correctedDelta = shortestDelta > 180 ? shortestDelta - 360 : shortestDelta;
+      const correctedDirection = direction + correctedDelta;
+
+      setCompassDirection(correctedDirection);
+    }, 32),
+    [],
+  );
+
+  const handleChange = useCallback((e: any) => {
+    clampPan(e);
+    updateCompassDirection(camera.position, ref.current.target);
+  }, []);
+
   const clampPan = useCallback(
     (e: any) => {
       _v.copy(e?.target.target);
       e?.target.target.clamp(minPan, maxPan);
       _v.sub(e?.target.target);
       camera.position.sub(_v);
-
-      // setCameraPosition(camera.position.clone());
-      // setCameraTarget(ref.current.target.clone());
     },
     [camera],
   );
@@ -124,7 +149,7 @@ const CameraControls = ({ position, target }: Props) => {
       minPolarAngle={minPolarAngle}
       zoomToCursor={isMapView}
       makeDefault
-      onChange={clampPan}
+      onChange={handleChange}
     />
   );
 };
